requirements.txt
# requirements.txt
# Alligator vs Crocodile Classification — Tilesview AI Interview Task
#
# Install all dependencies with:
#   pip install -r requirements.txt
#
# GPU users (CUDA 11.8):
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
#
# GPU users (CUDA 12.1):
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121

# ── Deep learning framework ──────────────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0

# ── Computer vision ───────────────────────────────────────────────────────────
Pillow>=9.4.0
opencv-python>=4.7.0

# ── Scientific computing ──────────────────────────────────────────────────────
numpy>=1.24.0
scikit-learn>=1.2.0

# ── Visualisation ─────────────────────────────────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0

# ── Dashboard ─────────────────────────────────────────────────────────────────
streamlit>=1.32.0
pandas>=1.5.0
